Modifications to fix bugs in "Equipments menu".
This commit has been squashed.
The modifications are
(1): An artificial limb and a costume were added for test.
(2): A bug was fixed: The button "back" had worked unexpectedlly.
(3): A bug was fixed: The animation2 had been played initially.
(4): The color of contents in the equipments list was improved.

BTW, when rebasing with squash in CUI, how can I use Japanese characters?